move +-*/ boilerplate to vect.py
